# swagger.json
{
  "swagger": "2.0",
  "paths": {
...
    "/api-key/generate": {
      "post": {
        "security": [
          {
            "api_key": [
              "[]"
            ]
          }
        ],
        "description": "\u003cb\u003eDescription\u003c/b\u003e\u003cbr\u003e\nGenerate API KEY\n\n\u003cbr\u003e\u003cb\u003e400: Bad Request\u003c/b\u003e\u003cbr\u003e\nErr400xBind_001 - Error while decoding request data to JSON - (could not decode parameters)\n\n\u003cbr\u003e\u003cb\u003e500: Internal Server Error\u003c/b\u003e\u003cbr\u003e\nErr500xAPIkey_001 - Error generating API Key - (server internal error)\nErr500xAPIkey_002 - Error storing API Key - (server internal error)\nErr500xTransaction_001 - Error while connecting to database - (server internal error)\nErr500xTransaction_002 - Error while committing all changes - (server internal error)",
        "tags": [
          "APIKey"
        ],
        "summary": "Generate API KEY ----- [Admin]",
        "operationId": "apiKeyGenerate",
        "parameters": [
          {
            "description": "API Key Data.",
            "name": "request",
            "in": "body",
            "schema": {
              "description": "API Key Data.",
              "type": "object",
              "$ref": "#/definitions/GenerateRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "API key response",
            "schema": {
              "$ref": "#/definitions/GenerateResponse"
            }
          },
          "default": {
            "description": "genericError",
            "schema": {
              "$ref": "#/definitions/genericError"
            }
          }
        }
      }
    },
...
    "/migrate": {
      "post": {
        "description": "\u003cb\u003eDescription\u003c/b\u003e\u003cbr\u003e\nLogin user using UDID, Credentials, Sequence Web3 Wallet\n\nIf \"type\" was set to \"udid\" - endpoint will expect fields 'udid' to be provided\nIf \"type\" was set to \"credentials\" - endpoint will expect fields 'email' and 'password' to be provided\nIf \"type\" was set to \"sequenceWallet\" - endpoint will expect fields 'walletAddress' and 'proof' to be provided\nIf \"type\" was set to \"sequenceWaasWallet\" - endpoint will expect fields 'token' to be provided\n\n\u003cbr\u003e\u003cb\u003e400: Bad Request\u003c/b\u003e\u003cbr\u003e\nErr400xBind_001 - Error while decoding request data to JSON - (could not decode parameters) \u003cbr\u003e\nErr400xBind_002 - Error while verifying if request data has all required fields - (not all fields found) \u003cbr\u003e\nErr400xAuthenticatorBind_001 - Error while decoding request data to JSON - (could not decode parameters) \u003cbr\u003e\nErr400xAuthenticatorBind_002 - Error while verifying if request data has all required fields - (not all fields found) \u003cbr\u003e\nErr400xAuthenticator_001 - Error while verifying if request data has all required fields - (unknown auth type) \u003cbr\u003e\n\n\u003cbr\u003e\u003cb\u003e403: Forbidden\u003c/b\u003e\u003cbr\u003e\nErr403xCredentialsAuth_001 - Error checking auth data - (invalid credentials) \u003cbr\u003e\n\n\u003cbr\u003e\u003cb\u003e409: Conflict\u003c/b\u003e\u003cbr\u003e\nErr409xAuthenticatorUser_001 - Error creating user - (auth option already in use) \u003cbr\u003e\n\n\u003cbr\u003e\u003cb\u003e500: Internal Server Error\u003c/b\u003e\u003cbr\u003e\nErr500xIssueToken_001 - Error while trying to issue access and refresh tokens - (server internal error) \u003cbr\u003e\nErr500xRoles_001 - Error while getting user roles - (server internal error) \u003cbr\u003e\nErr500xTransaction_001 - Error while connecting to database - (server internal error) \u003cbr\u003e\nErr500xTransaction_002 - Error while committing all changes - (server internal error) \u003cbr\u003e\nErr500xAuthenticatorTransaction_001 - Error while connecting to database - (server internal error) \u003cbr\u003e\nErr500xAuthenticatorTransaction_002 - Error while committing all changes - (server internal error) \u003cbr\u003e\nErr500xAuthenticatorUser_001 - Error while getting user - (server internal error) \u003cbr\u003e\nErr500xUDIDAuth_001 - Error getting user data - (server internal error) \u003cbr\u003e\nErr500xUDIDAuth_002 - Error checking auth data - (server internal error) \u003cbr\u003e\nErr500xUDIDAuth_003 - Error checking auth data - (server internal error) \u003cbr\u003e\nErr500xCredentialsAuth_001 - Error getting user data - (server internal error) \u003cbr\u003e\nErr500xCredentialsAuth_002 - Error checking auth data - (server internal error) \u003cbr\u003e\nErr500xCredentialsAuth_003 - Error checking auth data - (server internal error) \u003cbr\u003e\nErr500xCredentialsAuth_004 - Error while setting password - (server internal error) \u003cbr\u003e\nErr500xSequenceAuth_001 - Error validating wallet proof - (server internal error) \u003cbr\u003e\nErr500xSequenceAuth_002 - Error validating wallet proof - (proof is invalid) \u003cbr\u003e\nErr500xSequenceAuth_003 - Error getting user data - (server internal error) \u003cbr\u003e\nErr500xSequenceAuth_004 - Error checking auth data - (server internal error) \u003cbr\u003e\nErr500xSequenceAuth_005 - Error checking auth data - (server internal error) \u003cbr\u003e\nErr500xSequenceWaaSAuth_001 - Error validating token - (server internal error) \u003cbr\u003e\nErr500xSequenceWaaSAuth_002 - Error validating token - (token is invalid) \u003cbr\u003e\nErr500xSequenceWaaSAuth_003 - Error getting user data - (server internal error) \u003cbr\u003e\nErr500xSequenceWaaSAuth_004 - Error checking auth data - (server internal error) \u003cbr\u003e\nErr500xSequenceWaaSAuth_005\t- Error checking auth data - (server internal error) \u003cbr\u003e",
        "tags": [
          "Auth"
        ],
        "summary": "Login user",
        "operationId": "authSignin",
        "parameters": [
          {
            "description": "Sign In request data",
            "name": "request",
            "in": "body",
            "schema": {
              "description": "Sign In request data",
              "type": "object",
              "$ref": "#/definitions/SignInRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User is logged in successfully, returns tokens with userId",
            "schema": {
              "$ref": "#/definitions/SignInResponse"
            }
          },
          "default": {
            "description": "genericError",
            "schema": {
              "$ref": "#/definitions/genericError"
            }
          }
        }
      }
    },
...
    "/signin": {
      "post": {
        "description": "\u003cb\u003eDescription\u003c/b\u003e\u003cbr\u003e\nLogin user using UDID, Credentials, Sequence Web3 Wallet\n\nIf \"type\" was set to \"udid\" - endpoint will expect fields 'udid' to be provided\nIf \"type\" was set to \"credentials\" - endpoint will expect fields 'email' and 'password' to be provided\nIf \"type\" was set to \"sequenceWallet\" - endpoint will expect fields 'walletAddress' and 'proof' to be provided\nIf \"type\" was set to \"sequenceWaasWallet\" - endpoint will expect fields 'token' to be provided\n\n\u003cbr\u003e\u003cb\u003e400: Bad Request\u003c/b\u003e\u003cbr\u003e\nErr400xBind_001 - Error while decoding request data to JSON - (could not decode parameters) \u003cbr\u003e\nErr400xBind_002 - Error while verifying if request data has all required fields - (not all fields found) \u003cbr\u003e\nErr400xAuthenticatorBind_001 - Error while decoding request data to JSON - (could not decode parameters) \u003cbr\u003e\nErr400xAuthenticatorBind_002 - Error while verifying if request data has all required fields - (not all fields found) \u003cbr\u003e\nErr400xAuthenticator_001 - Error while verifying if request data has all required fields - (unknown auth type) \u003cbr\u003e\n\n\u003cbr\u003e\u003cb\u003e403: Forbidden\u003c/b\u003e\u003cbr\u003e\nErr403xCredentialsAuth_001 - Error checking auth data - (invalid credentials) \u003cbr\u003e\n\n\u003cbr\u003e\u003cb\u003e409: Conflict\u003c/b\u003e\u003cbr\u003e\nErr409xAuthenticatorUser_001 - Error creating user - (auth option already in use) \u003cbr\u003e\n\n\u003cbr\u003e\u003cb\u003e500: Internal Server Error\u003c/b\u003e\u003cbr\u003e\nErr500xIssueToken_001 - Error while trying to issue access and refresh tokens - (server internal error) \u003cbr\u003e\nErr500xRoles_001 - Error while getting user roles - (server internal error) \u003cbr\u003e\nErr500xTransaction_001 - Error while connecting to database - (server internal error) \u003cbr\u003e\nErr500xTransaction_002 - Error while committing all changes - (server internal error) \u003cbr\u003e\nErr500xAuthenticatorTransaction_001 - Error while connecting to database - (server internal error) \u003cbr\u003e\nErr500xAuthenticatorTransaction_002 - Error while committing all changes - (server internal error) \u003cbr\u003e\nErr500xAuthenticatorUser_001 - Error while getting user - (server internal error) \u003cbr\u003e\nErr500xUDIDAuth_001 - Error getting user data - (server internal error) \u003cbr\u003e\nErr500xUDIDAuth_002 - Error checking auth data - (server internal error) \u003cbr\u003e\nErr500xUDIDAuth_003 - Error checking auth data - (server internal error) \u003cbr\u003e\nErr500xCredentialsAuth_001 - Error getting user data - (server internal error) \u003cbr\u003e\nErr500xCredentialsAuth_002 - Error checking auth data - (server internal error) \u003cbr\u003e\nErr500xCredentialsAuth_003 - Error checking auth data - (server internal error) \u003cbr\u003e\nErr500xCredentialsAuth_004 - Error while setting password - (server internal error) \u003cbr\u003e\nErr500xSequenceAuth_001 - Error validating wallet proof - (server internal error) \u003cbr\u003e\nErr500xSequenceAuth_002 - Error validating wallet proof - (proof is invalid) \u003cbr\u003e\nErr500xSequenceAuth_003 - Error getting user data - (server internal error) \u003cbr\u003e\nErr500xSequenceAuth_004 - Error checking auth data - (server internal error) \u003cbr\u003e\nErr500xSequenceAuth_005 - Error checking auth data - (server internal error) \u003cbr\u003e\nErr500xSequenceWaaSAuth_001 - Error validating token - (server internal error) \u003cbr\u003e\nErr500xSequenceWaaSAuth_002 - Error validating token - (token is invalid) \u003cbr\u003e\nErr500xSequenceWaaSAuth_003 - Error getting user data - (server internal error) \u003cbr\u003e\nErr500xSequenceWaaSAuth_004 - Error checking auth data - (server internal error) \u003cbr\u003e\nErr500xSequenceWaaSAuth_005\t- Error checking auth data - (server internal error) \u003cbr\u003e",
        "tags": [
          "Auth"
        ],
        "summary": "Login user",
        "operationId": "authSignin",
        "parameters": [
          {
            "description": "Sign In request data",
            "name": "request",
            "in": "body",
            "schema": {
              "description": "Sign In request data",
              "type": "object",
              "$ref": "#/definitions/SignInRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User is logged in successfully, returns tokens with userId",
            "schema": {
              "$ref": "#/definitions/SignInResponse"
            }
          },
          "default": {
            "description": "genericError",
            "schema": {
              "$ref": "#/definitions/genericError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "DeviceID": {
      "type": "string",
      "x-go-package": "pr-reviewer/internal/model"
    },
...
    "GenerateRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "pr-reviewer/internal/api/v1/apiKey"
    },
    "GenerateResponse": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string",
          "x-go-name": "Key"
        }
      },
      "x-go-package": "pr-reviewer/internal/api/v1/apiKey"
    },
...
    "Role": {
      "description": "Role is a function a user can serve",
      "type": "string",
      "x-go-package": "pr-reviewer/internal/model"
    },
...
    "SignInRequest": {
      "type": "object",
      "properties": {
        "appVersion": {
          "type": "string",
          "x-go-name": "AppVersion"
        },
        "deviceID": {
          "$ref": "#/definitions/DeviceID"
        },
        "deviceType": {
          "type": "string",
          "x-go-name": "DeviceType"
        },
        "osVersion": {
          "type": "string",
          "x-go-name": "OSVersion"
        },
        "type": {
          "type": "string",
          "x-go-name": "Type"
        },
        "withPendingRequest": {
          "type": "boolean",
          "x-go-name": "WithPendingRequest"
        }
      },
      "x-go-name": "AddAuthRequest",
      "x-go-package": "pr-reviewer/internal/api/v1/auth"
    },
    "SignInResponse": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string",
          "x-go-name": "AccessToken"
        },
        "accessTokenExpiresAt": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "AccessTokenExpiresAt"
        },
        "pendingRequest": {
          "type": "boolean",
          "x-go-name": "PendingRequest"
        },
        "refreshToken": {
          "type": "string",
          "x-go-name": "RefreshToken"
        },
        "refreshTokenExpiresAt": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "RefreshTokenExpiresAt"
        },
        "roles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Role"
          },
          "x-go-name": "Roles"
        },
        "userID": {
          "$ref": "#/definitions/UserID"
        }
      },
      "x-go-name": "AddAuthResponse",
      "x-go-package": "pr-reviewer/internal/api/v1/auth"
    },
...
    "UserID": {
      "description": "UserID is identifier for User",
      "type": "string",
      "x-go-package": "pr-reviewer/internal/model"
    },
...
    "genericError": {
      "description": "GenericError - represent error structure for generic error (we need this to make all our error response the same)",
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "x-go-name": "ErrorCode"
        },
        "data": {
          "x-go-name": "Data"
        },
        "message": {
          "type": "string",
          "x-go-name": "ErrorMessage"
        }
      },
      "x-go-name": "GenericError",
      "x-go-package": "pr-reviewer/internal/api/middlewares"
    }
  }
}